Do not suppress Chmod on non-existent file
Currently fsnorify suppresses a Chmod event if the file does not exist
when event is received.

This prevents a possibility of fsnotify user to detect a situation
when an opened file is removed. In such case the Linux kernel sends
IN_ATTRIB event, as described in inotify(7) man page:

> IN_ATTRIB (*)
>        Metadata  changed—for example, permissions (e.g., chmod(2)),
>        timestamps (e.g., utimensat(2)), extended attributes  (setx‐
>        attr(2)), link count (since Linux 2.6.25; e.g., for the tar‐
>        get of link(2) and for unlink(2)), and user/group ID  (e.g.,
>        chown(2)).

To fix:
 * Modify the code to only suppress MODIFY and CREATE events.
 * Add a test case to verify Chmod event is delivered.

While at it, fix the comment in ignoreLinux() to use the up-to-date
terminology (event ops).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
(cherry picked from commit 486df7fb2b0ee2278338a51e01a0de12beb34c70)